% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox_model.R
\name{cox_model}
\alias{cox_model}
\title{cox models}
\usage{
cox_model(
  data,
  mortality,
  follow_up_time,
  independent_variables,
  covariates2,
  covariates3,
  weight = F,
  group = 1
)
}
\arguments{
\item{data}{data frame/weighted data.}

\item{mortality}{Is he/she in a state of death. 0/1.}

\item{follow_up_time}{follow-up time. numerical variable.}

\item{covariates2}{vector composed of covariates added to model2.}

\item{covariates3}{vector composed of covariates added to model3.}

\item{weight}{TRUE or FALSE, Weighted or not.}

\item{group}{If the independent variable is a categorical variable, please fill in its number of groups.}

\item{independent_rariables}{a vector composed of one or more independent variables. numerical variable.}
}
\description{
cox models
}
