% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_model.R
\name{logistic_model}
\alias{logistic_model}
\title{logistic models}
\usage{
logistic_model(
  data,
  dependent_variables,
  independent_variables,
  covariates2,
  covariates3,
  weight = F
)
}
\arguments{
\item{data}{data frame/weighted data.}

\item{covariates2}{vector composed of covariates added to model2.}

\item{covariates3}{vector composed of covariates added to model3.}

\item{weight}{TRUE or FALSE, Weighted or not.}

\item{dependent_rariables}{a vector composed of one or more dependent variables.}

\item{independent_rariables}{a vector composed of one or more independent variables.}
}
\description{
logistic models
}
