% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multmerge.R
\name{multmerge}
\alias{multmerge}
\title{Multiple merges}
\usage{
multmerge(data_list, by, all = F, all.x = all)
}
\arguments{
\item{data_list}{list of data frames to be merged}

\item{by}{vector indicating the variable(s) to merge by}

\item{all}{TRUE or FALSE}

\item{all.x}{TRUE or FALSE}
}
\description{
Multiple merges
}
