% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_count.R
\name{na_count}
\alias{na_count}
\title{Table NA counts of data.frames}
\usage{
na_count(data)
}
\arguments{
\item{data}{data frame}
}
\value{
A data frame with columns for Column names, NA_Count, and NA_Ratio
}
\description{
Table NA counts of data.frames
}
