% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_nas.R
\name{remove_nas}
\alias{remove_nas}
\title{Remove NAs and count remaining data}
\usage{
remove_nas(data, var_names)
}
\arguments{
\item{data}{data frame}

\item{var_names}{variable names for which to remove NAs}
}
\value{
A data frame with columns for Variable names, NA_Count, and Remaining_Count
}
\description{
Remove NAs and count remaining data
}
